bugfix : CurrencyManager 문제 해결

+++ Assets/01.Scripts/Ingame/Click/ClickInfo.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 public struct ClickInfo
 {
     public EClickType Type;
-    public int Damage;
+    public double Damage;
     public Vector2 Position;
 }
